+++ config/default.json
@@ -47,7 +47,8 @@
     "jobDomain": "RASTER",
     "ingestionNewJobType": "Ingestion_New",
     "ingestionUpdateJobType": "Ingestion_Update",
-    "ingestionSwapUpdateJobType": "Ingestion_Swap_Update"
+    "ingestionSwapUpdateJobType": "Ingestion_Swap_Update",
+    "initTaskType": "init"
   },
   "httpRetry": {
     "attempts": 5,
