--- src/common/enums.ts
@@ -1,3 +1,0 @@
-export enum TaskAction {
-  INIT = 'init',
-}
