+++ src/ingestion/schemas/updateMetadataSchema.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { z } from 'zod';
-import { CLASSIFICATION_REGEX } from '../../common/constants';
+import { CLASSIFICATION_REGEX } from './constants';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const createUpdateMetadataSchema = () => {
