+++ tests/integration/ingestion/helpers/ingestionRequestSender.ts
@@ -4,7 +4,7 @@ import { InputFiles, NewRasterLayer } from '@map-colonies/mc-model-types';
 export class IngestionRequestSender {
   public constructor(private readonly app: Express.Application) {}
 
-  public async validateIngestion(body: NewRasterLayer): Promise<supertest.Response> {
+  public async ingestNewLayer(body: NewRasterLayer): Promise<supertest.Response> {
     return supertest.agent(this.app).post('/ingestion').send(body).set('Content-Type', 'application/json');
   }
