+++ tests/mocks/configMock.ts
@@ -91,6 +91,7 @@ const registerDefaultConfig = (): void => {
       ingestionNewJobType: 'Ingestion_New',
       ingestionUpdateJobType: 'Ingestion_Update',
       ingestionSwapUpdateJobType: 'Ingestion_Swap_Update',
+      initTaskType: 'init',
     },
   };
